git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@489 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/TreeTest.cpp
@@ -39,7 +39,7 @@ bool TreeTest::isFreeTree(Graph *graph) {
 void TreeTest::makeRootedTree(Graph *graph, node root) {
   if (instance==0) instance=new TreeTest();
   graph->removeObserver(instance);
-  instance->resultsBuffer.erase((unsigned long)graph);
+  instance->resultsBuffer.erase((unsigned int)graph);
   if (!graph->isElement (root)) {
     cerr << "makeRootedTree:  Passed root is not element of graph" << endl;
     return;
